Update Back End

+++ src/main/java/com/postagem/atvpostagem/controller/PostagemController.java
@@ -2,6 +2,8 @@ package com.postagem.atvpostagem.controller;
 import java.util.List;
 import java.util.Optional;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,14 +15,16 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.postagem.atvpostagem.model.Postagem;
 import com.postagem.atvpostagem.repository.PostagemRepository;
 
 @RestController
 @RequestMapping("/postagens")
-@CrossOrigin("*")
+@CrossOrigin(value = "*", allowedHeaders = "*")
 public class PostagemController {
 
 	@Autowired
@@ -41,16 +45,22 @@ public class PostagemController {
 					return ResponseEntity.ok(repository.getByTituloContainingIgnoreCase(titulo));
 				}
 			@PostMapping
-				public ResponseEntity<Postagem> post (@RequestBody Postagem postagem){
+				public ResponseEntity<Postagem> post (@Valid @RequestBody Postagem postagem){
 					return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
 				}
-			@PutMapping
-				public ResponseEntity<Postagem> put (@RequestBody Postagem postagem){
-					return ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem));
+			@PutMapping 
+				public ResponseEntity<Postagem> put (@Valid @RequestBody Postagem postagem){
+				return repository.findById(postagem.getId())
+				        .map(resp -> ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem)))
+				        .orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 				}
+			@ResponseStatus(HttpStatus.NO_CONTENT)
 			@DeleteMapping("/{id}")
 				public void delete(@PathVariable Long id) {
-					repository.deleteById(id);
+			     Optional<Postagem> post = repository.findById(id);
+			        if(post.isEmpty())
+				        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+			        repository.deleteById(id);
 				}
 	}
